Added update to terminal interface.
Refactored run to a new file (run.py), added config.py with version and title. Added update to terminal interface with version number and copyright.
