+++ pyscrapetrain/config.py
@@ -1,0 +1,13 @@
+__version__ = "0.1.2"
+
+__title__ = r"""
+                                                              ______                              
+                                                             /\__  _\               __            
+ _____   __  __    ____    ___   _ __    __     _____      __\/_/\ \/ _ __    __   /\_\    ___    
+/\ '__`\/\ \/\ \  /',__\  /'___\/\`'__\/'__`\  /\ '__`\  /'__`\ \ \ \/\`'__\/'__`\ \/\ \ /' _ `\  
+\ \ \L\ \ \ \_\ \/\__, `\/\ \__/\ \ \//\ \L\.\_\ \ \L\ \/\  __/  \ \ \ \ \//\ \L\.\_\ \ \/\ \/\ \ 
+ \ \ ,__/\/`____ \/\____/\ \____\ \ \_\\ \__/.\_\\ \ ,__/\ \____\  \ \_\ \_\\ \__/.\_\\ \_\ \_\ \_\
+  \ \ \/  `/___/> \/___/  \/____/ \/_/ \/__/\/_/ \ \ \/  \/____/   \/_/\/_/ \/__/\/_/ \/_/\/_/\/_/
+   \ \_\     /\___/                               \ \_\                                           
+    \/_/     \/__/                                 \/_/                                           
+"""
